Prepare release for tizen.org
Changes:
 * Reset manifest.
 * Security policy for security-server sockets has been removed.
 * Minor fixes required by g++ 4.8.
 * Removed unused code.
 * Version update.
 * Fix conflicts in manifest file.

[Issue#]   SSDWSSP-816
[Bug]      N/A
[Cause]    N/A
[Solution] N/A

[Verification] N/A

+++ src/server/service/password-file.cpp
@@ -123,10 +123,10 @@ namespace SecurityServer
         unsigned int algorithm;
         Deserialization::Deserialize(stream, algorithm);
         switch (algorithm) {
-            case IPassword::PasswordType::NONE:
+            case (unsigned int)IPassword::PasswordType::NONE:
                 ptr.reset(new NoPassword());
                 break;
-            case IPassword::PasswordType::SHA256:
+            case (unsigned int)IPassword::PasswordType::SHA256:
                 ptr.reset(new SHA256Password(stream));
                 break;
             default:
